tmux: capture-pane
